+++ example/lib/page/index.dart
@@ -1,16 +1,18 @@
 import 'package:flutter/material.dart';
-import 'package:ijkplayer_example/i18n/i18n.dart';
-import 'package:ijkplayer_example/page/asset_page.dart';
-import 'package:ijkplayer_example/page/controller_stream_use.dart';
-import 'package:ijkplayer_example/page/dialog_video_page.dart';
-import 'package:ijkplayer_example/page/full_screen.dart';
-import 'package:ijkplayer_example/page/gallery_page.dart';
-import 'package:ijkplayer_example/page/ijk_status_page.dart';
-import 'package:ijkplayer_example/page/in_overlay_page.dart';
-import 'package:ijkplayer_example/page/network.dart';
 import 'package:ijkplayer_example/page/paging_page.dart';
 import 'package:ijkplayer_example/page/screen_shot_page.dart';
 import 'package:ijkplayer_example/page/video_list.dart';
+import '../i18n/i18n.dart';
+
+import 'asset_page.dart';
+import 'controller_stream_use.dart';
+import 'custom_ijk_opt_page.dart';
+import 'dialog_video_page.dart';
+import 'full_screen.dart';
+import 'gallery_page.dart';
+import 'ijk_status_page.dart';
+import 'in_overlay_page.dart';
+import 'network.dart';
 
 class IndexPage extends StatefulWidget {
   @override
@@ -38,6 +40,7 @@ class _IndexPageState extends State<IndexPage> {
           buildButton(currentI18n.screenshotTitle, ScreenShotPage()),
           buildButton(currentI18n.overlayPageTitle, InOverlayPage()),
           buildButton(currentI18n.ijkStatusTitle, IjkStatusPage()),
+          buildButton(currentI18n.customOption, CustomIjkOptionPage()),
         ],
       ),
     );